+++ src/main/java/com/project/repository/FlightRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.project.dto.UserInput;
 import com.project.model.Flight;
 
-public interface FlightRepository extends JpaRepository<com.project.model.Flight, Integer> {
+public interface FlightRepository extends JpaRepository<com.project.model.Flight, Long> {
